Add logger and initialize lottoDataBase (functional but need to work on Hive)
